made recommendationscreen dynamic; uses one screen to display the diff categories
