--- src/screens/RecommendationScreens/ActivitiesRecs.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import {View} from 'react-native';
-
-const ActivitiesRecs = ({navigation}) => {
-  return <View />;
-};
-export default ActivitiesRecs;
