--- src/screens/RecommendationScreens/RestaurantRecs.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import {View} from 'react-native';
-
-const RestaurantRecs = ({navigation}) => {
-  return <View />;
-};
-export default RestaurantRecs;
